ensure STATUS_STARTING is sent as first step in startup
Since TorService should be running whenever tor is running,
TorService.onCreate() is the very first step in the startup procedure, so
that is where STATUS_STARTING should be announced to make sure that the UI
reflects the TorService state as close as possible.

But STATUS_STARTING should not be sent if the tor daemon is already running
so this checks for a running tor daemon as the very first step of starting
TorService.