use global status var to get tor status instantly in Activities
Since TorService is a regular Service and not an IntentService or remote
Service, it runs as part of the same process as the rest of the app.  That
means that OrbotMainActivity can access static variables set in TorService.
Since there is only a single TorService by definition, the status can be
stored

This drastically simplifies the startup process since OrbotMainActivity no
longer needs to request status info, then wait for a reply.