force all Intents coming into TorService into a boundless queue
Now that any app can send a start request directly to TorService, and there
are other requests that can come in as well, the start/stop procedure needs
to be protected by a queue.  If TorService received a bunch of start
commands in a short amount of time, they would wreak havoc if they all ran
concurrently.

That said, some of the stuff in IncomingIntentRouter would work better not
behind a queue, but I haven't addressed that yet.  For example, it is
perfectly fine to execute many ACTION_STATUS Intents at the same time, as
well as when TorService is STATUS_STARTING.