move torUpgradeAndConfig() to IncomingIntentRouter so it runs in the thread
Now that all incoming Intents are queued, we can move torUpgradeAndConfig()
to IncomingIntentRouter, where it will also be run in the thread.  That'll
make Orbot start up faster when it is running the upgrade procedure.

This also moves it to be behind the queue of incoming Intents, so that it
guarantees that the daemons are unpacked before anything tries to run.